Move internal/hub to pkg/hub
