+++ internal/commands/root.go
@@ -33,7 +33,7 @@ import (
 	"github.com/docker/hub-tool/internal/commands/tag"
 	"github.com/docker/hub-tool/internal/commands/token"
 	"github.com/docker/hub-tool/internal/credentials"
-	"github.com/docker/hub-tool/internal/hub"
+	"github.com/docker/hub-tool/pkg/hub"
 	"github.com/docker/hub-tool/internal/login"
 )
