+++ internal/commands/token/inspect.go
@@ -30,7 +30,7 @@ import (
 
 	"github.com/docker/hub-tool/internal/ansi"
 	"github.com/docker/hub-tool/internal/format"
-	"github.com/docker/hub-tool/internal/hub"
+	"github.com/docker/hub-tool/pkg/hub"
 	"github.com/docker/hub-tool/internal/metrics"
 )
