+++ internal/login/login.go
@@ -33,7 +33,7 @@ import (
 	"github.com/docker/hub-tool/internal/ansi"
 	"github.com/docker/hub-tool/internal/credentials"
 	"github.com/docker/hub-tool/internal/errdef"
-	"github.com/docker/hub-tool/internal/hub"
+	"github.com/docker/hub-tool/pkg/hub"
 )
 
 // RunLogin logs the user and asks for the 2FA code if needed
