+++ main.go
@@ -30,7 +30,7 @@ import (
 
 	"github.com/docker/hub-tool/internal/commands"
 	"github.com/docker/hub-tool/internal/credentials"
-	"github.com/docker/hub-tool/internal/hub"
+	"github.com/docker/hub-tool/pkg/hub"
 )
 
 func main() {
