+++ pkg/hub/client_test.go
