+++ pkg/hub/error.go
